EVENT now works correctly(may have some bugs?). Fixed a bug in export.
I've added two methods in RUN.java. They are raceInProgress() and
getQueue(). getQueue() will add all existing racers in queue of that run
to a arraylist and return it. Event will add each racer from the
arraylist to the new run. raceInProgress() will check if there are any
racers running or completed, returning true which does not allow the
user to change the event type.